feat-40: debug rooms controller

+++ src/main/kotlin/com/egecube/eduplatform/quizGame/rooms/domain/PlayerInRoom.kt
@@ -2,5 +2,5 @@ package com.egecube.eduplatform.quizGame.rooms.domain
 
 data class PlayerInRoom(
     val userId: Long,
-    val roomId: Long
+    val roomId: Int
 )
